Housekeeping for less complexity and more readability
